Implemented algorithms in precision form
This commit branches the current code in order to explore
the feasibility of algorithms derived from the precision
construction of the probability distributions. In fact,
the algorithms are much closer to IRLS in this form.

One interesting note: in this form, extended/full VRLS
is still sensitive to its hyperparameters beta_tau and
beta_xi. The original construction may yet be more
effective at "desensitizing"...