Added RMSE-form solvers
This branch implements IRLS, VRLS and GRLS solvers
that were derived from the "root-mean-square error"
construction of the probability distributions.

While it appeared initially that the solvers in this
form were less sensitive to their hyperparameter
settings, this was not the case.

Frankly, the equations in this form are considerably
more awkward than in the precision form...