Fixed a bug in IRLS-EM and IRLS-MAP
These two algorithms had x-updates that were missing
a factor of two in front of the weight matrix. This
has been corrected.

Considering the fact that multiple subtle bugs have
been found in several solvers now, it will be very
necessary to fully review the derivations of each
algorithm.

A compendium of solver derivations is in progress.